Good output for mamba-370m

+++ examples/offline_inference.py
@@ -1,4 +1,5 @@
 from vllm import LLM, SamplingParams
+import torch
 
 # Sample prompts.
 prompts = [
@@ -11,7 +12,7 @@ prompts = [
 sampling_params = SamplingParams(temperature=0.0, top_p=0.95)
 
 # Create an LLM.
-llm = LLM(model="state-spaces/mamba-370m-hf")
+llm = LLM(model="state-spaces/mamba-370m-hf", dtype=torch.float32)
 # Generate texts from the prompts. The output is a list of RequestOutput objects
 # that contain the prompt, generated text, and other information.
 outputs = llm.generate(prompts, sampling_params)
